Remove gulp with simpler npm tasks
This removes `gulp` and provides a simpler `karma` and `webpack`,
`webpack-watch` npm scripts.

For actual development, users will need the Rails stack anyway, so it
won't make sense to keep a separate set of dependencies up to date.

* Instead of `gulp webpack`, use `npm run webpack`.
* Instead of `gulp watch`, use `npm run webpack-watch`, which is *drastically* faster than gulp watch-polling.
* Instead of `gulp tests:karma` use `npm run karma`, or just `npm test`.

This also removes all traces of express and restores the Rails
styleguide. It is not whitelisted for precompilation and is only
available in the context of the development mode.

To access the styleguide, use `/assets/styleguide.html`.
