+++ .gitignore
@@ -81,6 +81,7 @@
 .rspec
 # coverage in plugins
 /lib/plugins/*/coverage
-
 # asset cache
 /.sass-cache/
+# Frontend debug log
+/frontend/npm-debug.log*
