+++ Procfile.dev
@@ -1,3 +1,3 @@
-web: bundle exec unicorn --config-file config/unicorn.rb --host ${HOST:="127.0.0.1"} --port ${PORT:="8080"} --env ${RAILS_ENV:="development"}
-assets: cd frontend && $(npm bin)/webpack --colors --watch --progress
+web: bundle exec rails server -p ${PORT:="8080"} -b ${HOST:="127.0.0.1"} --env ${RAILS_ENV:="development"}
+assets: cd frontend && npm run webpack-watch
 worker: bundle exec rake jobs:work
