+++ app/assets/stylesheets/styleguide.js
@@ -28,3 +28,19 @@
 
 //= require ../javascripts/bundles/openproject-global
 //= require ../javascripts/tooltips
+
+angular.module('openproject-style-guide', ['ui.select', 'ngSanitize']);
+
+// Add uiComponents to the styleguide.
+// In order to be able to do that, we have to mock some
+// services that directives in uiComponents rely on.
+angular.module('openproject-style-guide')
+  .service('ActivityService', function() {} )
+  .service('ConfigurationService', function() {} )
+  .service('AutoCompleteHelper', function() {} )
+  .service('NotificationsService', function() {
+    return {
+      addError: function() {},
+      addSuccess: function() {}
+    };
+}).requires.push('openproject.uiComponents');
