+++ frontend/app/global.js
@@ -34,6 +34,8 @@
 // See: https://github.com/webpack/style-loader/issues/31
 require('polyfill-function-prototype-bind');
 
+require('i18n');
+
 require('jquery');
 require('jquery-migrate/jquery-migrate');
 require('jquery-ujs');
