+++ frontend/doc/BUILD.md
@@ -1,11 +1,9 @@
 Building
 ========
 
-__Note__: All tasks involved are found in `./frontend/gulpfile.js`
+All builds are put into Rails' asset pipeline. The actual build, i.e. merging all components together is done via [`webpack`](https://github.com/webpack/webpack).
 
-All builds are put into Rails' asset pipeline. Builds are run via [`gulp`](http://gulpjs.com/). The actual build, i.e. merging all components together is done via [`webpack`](https://github.com/webpack/webpack).
-
-It __is important to note__ that OpenProject currently still relies on the asset pipeline to serve the assets. Minification is done via `rake assets:precompile` and does __not__ happen in the `gulp` based pipeline.
+It __is important to note__ that OpenProject currently still relies on the asset pipeline to serve the assets. Minification is __not__ done as part of webpack.
 
 Two types of builds are performed, the first one is the OpenProject Standalone JS, which is not required by the Rails views per se. The second one is a bundle of global dependencies, which are necessary for the Rails views to run properly.
 
@@ -15,7 +13,7 @@ The configuration for building both global and standalone JS is found in `./fron
 
 The resulting output of this buildstep can be found in `./app/assets/javascripts/bundles/openproject-core-app.js`
 
-This is done via `gulp webpack` (see `gulpfile.js`, Line 63 ff.). The actual entrypoint for this is `./frontend/app/openproject-app.js`.
+This is done via `npm run webpack` (see `frontend/package.json`). The actual entrypoint for this is `./frontend/app/openproject-app.js`.
 
 It contains only the the JavaScript needed for the AngularJS based part of the codebase (and `lodash`).
 
@@ -23,13 +21,11 @@ It contains only the the JavaScript needed for the AngularJS based part of the c
 
 The resulting output of this buildstep can be found in `./app/assets/javascripts/bundles/openproject-global.js`. 
 
-This is done via `gulp webpack` (see `gulpfile.js`, Line 63 ff.). The actual entrypoint for this is `./frontend/app/global.js`
+This is also built with `npm run webpack`. The actual entrypoint for this is `./frontend/app/global.js`
 
 It contains all of the JavaScript necessary for the rails based views, like parts of `jQuery.ui`, but also `angular` itself.
 
 ## Building Sass
-
-Sass files are built via `gulp sass`, which handles the main Sass file from the Rails stack at `./app/assets/stylesheets/defaults.css.sass`, performs transformations on it and outputs the result to `./frontend/public/assets`. It __is important__ to note, that this serves __not__ the compilation of Sass for production purposes but is mostly for the availability of the CSS for the Living Styleguide. 
 
 The Sass files in the rails stack are handled as one would expect: They are precompiled into one `default-*.css` for production and are reloaded on demand during development. The manifest for this can be found in `./app/assets/stylesheets/defaults.css.sass`.
 
@@ -37,10 +33,8 @@ The Sass files in the rails stack are handled as one would expect: They are prec
 
 OpenProject relies on the [Foundation for Apps Framework](http://foundation.zurb.com/apps), as well as one the [Bourbon Mixin Library](http://bourbon.io/). 
 
-They are provided via LoadPath manipulation in the `gulp` based pipeline.The two frameworks are included as  `bower` components (see `./frontend/bower.json`).
-
 On the Rails side, both frameworks are included as gems - see the `./Gemfile` - and plugged directly into Rails' asset pipeline.
 
-## Misc Tasks
+## Living Styleguide
 
-The build pipeline is also responsible for building the [Living Styleguide](https://github.com/livingstyleguide/livingstyleguide), relying heavily on the duplicated functionality of the `gulp` tasks revolving around Sass compilation
+The styleguide is rendered as part of the Rails stack at `/assets/styleguide.html`. Please see [their GitHub](https://github.com/livingstyleguide/livingstyleguide) for more information.
