+++ frontend/scripts/karma-runner.js
@@ -1,0 +1,84 @@
+#!/usr/bin/env node
+
+//-- copyright
+// OpenProject is a project management system.
+// Copyright (C) 2012-2015 the OpenProject Foundation (OPF)
+//
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU General Public License version 3.
+//
+// OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+// Copyright (C) 2006-2013 Jean-Philippe Lang
+// Copyright (C) 2010-2013 the ChiliProject Team
+//
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU General Public License
+// as published by the Free Software Foundation; either version 2
+// of the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+//
+// See doc/COPYRIGHT.rdoc for more details.
+//++
+
+var path = require('path');
+var tsproject = require('tsproject');
+var karma = require('karma');
+var fs = require('fs');
+
+function deleteFolderRecursive(path) {
+  if( fs.existsSync(path) ) {
+    fs.readdirSync(path).forEach(function(file){
+      var curPath = path + "/" + file;
+
+      if(fs.lstatSync(curPath).isDirectory()) {
+        deleteFolderRecursive(curPath);
+
+      } else {
+        fs.unlinkSync(curPath);
+      }
+    });
+
+    fs.rmdirSync(path);
+  }
+}
+
+// Clean old TS build dir
+var frontend_path = path.resolve(__dirname, '..');
+var tsOutDir = path.resolve(frontend_path, 'tests/unit/tests/typescript');
+deleteFolderRecursive(tsOutDir);
+
+// Compile TS
+tsproject.src(path.resolve(frontend_path, 'tsconfig.test.json'), {
+  compilerOptions: {
+    outDir: tsOutDir
+  }
+});
+
+// Run karma
+karma.server.start(
+  {
+    configFile: path.resolve(frontend_path, 'karma.conf.js'),
+    singleRun: true
+  },
+  function (exitCode) {
+    if(exitCode === 0) {
+      console.info('No tests have failed');
+      console.info('Files generated by tsc were deleted.');
+      deleteFolderRecursive(tsOutDir);
+    }
+    else {
+      console.warn('Tests have failed');
+      console.info('Files generated by tsc can be found in: ' + tsOutDir);
+    }
+
+    process.exit(exitCode);
+  }
+);
