+++ package.json
@@ -4,7 +4,9 @@
   "scripts": {
     "postinstall": "cd frontend && npm install",
     "test": "cd frontend && npm test && cd ..",
-    "karma": "cd frontend && npm run karma && cd .."
+    "karma": "cd frontend && npm run karma && cd ..",
+    "webpack": "cd frontend && npm run webpack && cd ..",
+    "webpack-watch": "cd frontend && npm run webpack-watch"
   },
   "private": true
 }
